backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-CORS==3.0.10
Flask-Migrate==4.0.4
python-dotenv==1.0.0
twilio==8.3.0
geopy==2.3.0
PyJWT==2.7.0
cryptography==41.0.3
Flask-Limiter==3.5.0
sentry-sdk==1.32.0
gunicorn==21.2.0